+++ NEWS.md
@@ -1,6 +1,11 @@
-## version 0.1.0 [2020-10-06]
+## version 0.1.0 [2021-10-06]
 
 * Implemented speed permutation-based statistical tests (one sample and two samples)
 
+## version 0.1.1 [2021-11-05]
 
-
+* Add pecora-package.md
+* Changed data matrix X dimension
+* Changed documentation and rand arguments
+* Deleted 
+* Created separted function to compute p-values